requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.8.3
gspread==6.1.2
google-auth==2.29.0
python-dateutil==2.9.0
